Make unit test more accurate
The goal of the test was to show that Sun cards were always played first, regardless of Player behavior.

However, the Sun card was placed such that the Player would always select it, meaning we couldn't tell if the State was actually doing the right thing.

Fix: move the Sun card to where the Player would NOT select it to show that the State is selecting it.