Support for multiple owls
Owls' positions are stored in a list on GameBoard. Plays must now include which
position they are being played on. Position is used instead of owl index since
owls are fungible.

IPlayer's SelectCardToPlay renamed to FormulatePlay to emphasize that a position
is also part of the required return value. In order to preserve the intent of
GameStateTests.ShouldWinGameWhenOwlsReachNest, the LeastRecentCardPlayer selects
the owl that is furtherest from the start point.