GameState is only state. Game is pure functions.
In order to help address several future concerns, we want to have a (vaguely) immutable representation of the state of the board. In addition, we want to be able to produce the next state given a play. This is important so that algorithms that wish to tentatively look ahead assuming they've made a particular play may do so without altering the game in progress.