GameState equality does value comparisons instead of reference
Since we will need to be able to detect duplicate states for several
algorithms, we have implemented a way to compare states to each other. We made
the decision to include the deck in the comparison which may need additional
attention if/when we implement agents without knowledge of the deck.