GetHashCode, ==, and != implementations for GameState
Since we are overriding Equals on the GameState objects, it is highly
recommended that we also implement GetHashCode, ==, and != in order to ensure
other data structures treat these objects correctly and efficiently.

There may be a way to DRY up these methods using a static helper class and/or an abstract parent class.