SearchNode and GreedyBestFirstAgent
The primary goal here was to implement GreedyBestFirstSearch. To do that, we
needed a way to store our game tree traversal, hence SearchNode. In order to
have a bit of added safety around gameplay, a StateMismatchException will be
thrown if the GreedyBestFirstAgent is ever asked to make a play that doesn't
correspond to what it believes the state of the board should be.