Adding support for expectimax in game state and deck
GameState and Deck are going to require some changes to allow us to
implement an expectimax agent.  With expectimax the SearchNodes need a
ChanceNode type, but this would only work in HootOwlHoot if we also are
able to separate a successor action into multiple phases.  All other
Agents will continue to use normal Successor functionality, but the
expectimax will separate things into multiple phases.

Part one of this is to:
    - Decks can provide a list of probability of all cards left in them
    - Decks can force draw a card type and the probability of drawing
    that card
    - GameState can iteratively take a turn in multiple phases